Updates to nekRS-ML benchmark (#8)
Remove nekRS-ML as a git submodule to the repository and instead points to the new nekRS-ML repo and the specific branch being used for the ALCF-4 benchmark.

It also updates the build and run instructions, and the required links.

Finally, FOM_shoot is updated to include the time step size ratio between the GNN and nekRS. This changes the value of the FOM, but not the spirit of it.